chore: Convert sheets on `/now-playing` route to use True Sheet
- The "Playback Options" sheet is still using `react-native-actions-sheet` due to some gesture problems.
- Ensure that `queueList` is populated when we open the "Upcoming Tracks" sheet.
- Fix slider "labels" randomly doubling in height.
- Support using ref to present sheet with old library.
  - Also updated backdrop opacity to match.
