+++ mobile/src/app/now-playing.tsx
@@ -17,8 +17,9 @@ import { useMusicStore } from "~/modules/media/services/Music";
 import { MusicControls } from "~/modules/media/services/Playback";
 import { Router } from "~/services/NavigationStore";
 import { useUserPreferencesStore } from "~/services/UserPreferences";
-import { useSeekStore } from "~/screens/NowPlaying/SeekService";
+import { useSeekStore } from "~/screens/NowPlaying/helpers/SeekService";
 import { NowPlayingArtwork } from "~/screens/NowPlaying/Artwork";
+import { NowPlayingSheets } from "~/screens/NowPlaying/Sheets";
 
 import { mutateGuard } from "~/lib/react-query";
 import { cn } from "~/lib/style";
@@ -27,6 +28,8 @@ import { Marquee } from "~/components/Containment/Marquee";
 import { SafeContainer } from "~/components/Containment/SafeContainer";
 import { IconButton } from "~/components/Form/Button";
 import { Slider } from "~/components/Form/Slider";
+import { useSheetRef } from "~/components/Sheet";
+import { useLegacySheetRef } from "~/components/Sheet/Legacy";
 import { Back } from "~/components/Transition/Back";
 import { StyledText } from "~/components/Typography/StyledText";
 import {
@@ -196,30 +199,40 @@ function PlaybackControls() {
 /** Actions rendered on the bottom of the screen. */
 function BottomAppBar() {
   const { t } = useTranslation();
+  const playbackOptionsSheetRef = useLegacySheetRef();
+  const upcomingTracksSheetRef = useSheetRef();
+
   return (
-    <View className="flex-row items-center justify-between gap-4 p-4">
-      <BackButton />
-      <View className="flex-row items-center gap-4">
-        <IconButton
-          kind="ripple"
-          accessibilityLabel={t("feat.playback.extra.options")}
-          onPress={() => SheetManager.show("PlaybackOptionsSheet")}
-          rippleRadius={24}
-          className="p-2"
-        >
-          <InstantMix size={32} />
-        </IconButton>
-        <IconButton
-          kind="ripple"
-          accessibilityLabel={t("term.upcoming")}
-          onPress={() => SheetManager.show("TrackUpcomingSheet")}
-          rippleRadius={24}
-          className="p-2"
-        >
-          <LibraryMusic size={32} />
-        </IconButton>
+    <>
+      <View className="flex-row items-center justify-between gap-4 p-4">
+        <BackButton />
+        <View className="flex-row items-center gap-4">
+          <IconButton
+            kind="ripple"
+            accessibilityLabel={t("feat.playback.extra.options")}
+            onPress={() => playbackOptionsSheetRef.current?.show()}
+            rippleRadius={24}
+            className="p-2"
+          >
+            <InstantMix size={32} />
+          </IconButton>
+          <IconButton
+            kind="ripple"
+            accessibilityLabel={t("term.upcoming")}
+            onPress={() => upcomingTracksSheetRef.current?.present()}
+            rippleRadius={24}
+            className="p-2"
+          >
+            <LibraryMusic size={32} />
+          </IconButton>
+        </View>
       </View>
-    </View>
+
+      <NowPlayingSheets
+        playbackOptionsRef={playbackOptionsSheetRef}
+        upcomingTracksRef={upcomingTracksSheetRef}
+      />
+    </>
   );
 }
